+++ src/main/java/com/gymmanager/gym_manager/repository/AsistenciaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gymmanager.gym_manager.entity.Asistencia;
 
-public interface AsistenciaRepository extends JpaRepository<Asistencia, Long> {
+public interface AsistenciaRepository extends JpaRepository<Asistencia, Integer> {
 
 }
